04:36 2021-5-7 금
UI부분 대충 작동 완료.
씬 체인지 작동 완료.
내일 꼭 네비매쉬 끝내기

+++ Engine/Cpp/Anim_Controller.cpp
@@ -192,7 +192,7 @@ float Anim_Controller::Get_CurFrame() const
 
 float Anim_Controller::Get_MaxFrame() const
 {
-	return m_dMaxKeyFrame;
+	return (float)m_dMaxKeyFrame;
 }
 
 bool Anim_Controller::Get_End()
